requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
bcrypt==4.3.0
gunicorn==22.0.0
python-dotenv==1.0.1
mysqlclient==2.2.4